📝 (Docker) Use same Dockerfile in lectures 2 and 3
Previously the Lecture 2 Dockerfile shown was more advanced and more "correct", however, we then use a different Dockerfile in practice in Lecture 3.

I think it's better to use the same, slightly less good Dockerfile, throughout the section, and then introduce the "better" Dockerfile when we deploy the app and use `gunicorn`.