🐛 fix(Item Update): Fix `ItemUpdateSchema` by adding `store_id`
This is not necessary just for updating, but it is necessary for creating items. Since the item update endpoint may update or create items, the `store_id` is added as an optional field to the `ItemUpdateSchema` so that it can be accepted by the API when a client wants to create an item.